requirements.txt
fastapi
uvicorn
pandas
paho-mqtt
motor
numpy
tensorflow
keras
